chore[8.10]: rename queries folder to graphql
