+++ part8/frontend/src/graphql/queries.js
